Cleanup project  - Deprecate unused classes  - Move private classes to inner classes
